Move serializers to a separate package
