+++ core/common/src/DateTimePeriod.kt
@@ -5,80 +5,12 @@
 
 package kotlinx.datetime
 
+import kotlinx.datetime.serializers.DatePeriodISO8601Serializer
+import kotlinx.datetime.serializers.DateTimePeriodISO8601Serializer
 import kotlin.math.*
 import kotlin.time.Duration
 import kotlin.time.ExperimentalTime
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-
-object DateTimePeriodComponentSerializer: KSerializer<DateTimePeriod> {
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("DateTimePeriod") {
-            element<Int>("years", isOptional = true)
-            element<Int>("months", isOptional = true)
-            element<Int>("days", isOptional = true)
-            element<Int>("hours", isOptional = true)
-            element<Int>("minutes", isOptional = true)
-            element<Int>("seconds", isOptional = true)
-            element<Long>("nanoseconds", isOptional = true)
-        }
-
-    override fun deserialize(decoder: Decoder): DateTimePeriod =
-        decoder.decodeStructure(descriptor) {
-            var years = 0
-            var months = 0
-            var days = 0
-            var hours = 0
-            var minutes = 0
-            var seconds = 0
-            var nanoseconds = 0L
-            loop@while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> years = decodeIntElement(descriptor, 0)
-                    1 -> months = decodeIntElement(descriptor, 1)
-                    2 -> days = decodeIntElement(descriptor, 2)
-                    3 -> hours = decodeIntElement(descriptor, 3)
-                    4 -> minutes = decodeIntElement(descriptor, 4)
-                    5 -> seconds = decodeIntElement(descriptor, 5)
-                    6 -> nanoseconds = decodeLongElement(descriptor, 6)
-                    CompositeDecoder.DECODE_DONE -> break@loop // https://youtrack.jetbrains.com/issue/KT-42262
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            DateTimePeriod(years, months, days, hours, minutes, seconds, nanoseconds)
-        }
-
-    override fun serialize(encoder: Encoder, value: DateTimePeriod) {
-        encoder.encodeStructure(descriptor) {
-            with(value) {
-                if (years != 0) encodeIntElement(descriptor, 0, years)
-                if (months != 0) encodeIntElement(descriptor, 1, months)
-                if (days != 0) encodeIntElement(descriptor, 2, days)
-                if (hours != 0) encodeIntElement(descriptor, 3, hours)
-                if (minutes != 0) encodeIntElement(descriptor, 4, minutes)
-                if (seconds != 0) encodeIntElement(descriptor, 5, seconds)
-                if (nanoseconds != 0) encodeLongElement(descriptor, 6, value.nanoseconds.toLong())
-            }
-        }
-    }
-
-}
-
-object DateTimePeriodISO8601Serializer: KSerializer<DateTimePeriod> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("DateTimePeriod", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): DateTimePeriod =
-        DateTimePeriod.parse(decoder.decodeString())
-
-    override fun serialize(encoder: Encoder, value: DateTimePeriod) {
-        encoder.encodeString(value.toString())
-    }
-
-}
+import kotlinx.serialization.Serializable
 
 @Serializable(with = DateTimePeriodISO8601Serializer::class)
 // TODO: could be error-prone without explicitly named params
@@ -304,78 +236,6 @@ sealed class DateTimePeriod {
 }
 
 public fun String.toDateTimePeriod(): DateTimePeriod = DateTimePeriod.parse(this)
-
-object DatePeriodComponentSerializer: KSerializer<DatePeriod> {
-
-    private fun unexpectedNonzero(fieldName: String, value: Long) {
-        if (value != 0L) {
-            throw SerializationException("expected field '$fieldName' to be zero, but was $value")
-        }
-    }
-
-    private fun unexpectedNonzero(fieldName: String, value: Int) = unexpectedNonzero(fieldName, value.toLong())
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("DatePeriod") {
-            element<Int>("years", isOptional = true)
-            element<Int>("months", isOptional = true)
-            element<Int>("days", isOptional = true)
-            element<Int>("hours", isOptional = true)
-            element<Int>("minutes", isOptional = true)
-            element<Long>("seconds", isOptional = true)
-            element<Long>("nanoseconds", isOptional = true)
-        }
-
-    override fun deserialize(decoder: Decoder): DatePeriod =
-        decoder.decodeStructure(descriptor) {
-            var years = 0
-            var months = 0
-            var days = 0
-            loop@while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> years = decodeIntElement(descriptor, 0)
-                    1 -> months = decodeIntElement(descriptor, 1)
-                    2 -> days = decodeIntElement(descriptor, 2)
-                    3 -> unexpectedNonzero("hours", decodeIntElement(descriptor, 3))
-                    4 -> unexpectedNonzero("minutes", decodeIntElement(descriptor, 4))
-                    5 -> unexpectedNonzero("seconds", decodeLongElement(descriptor, 5))
-                    6 -> unexpectedNonzero("nanoseconds", decodeLongElement(descriptor, 6))
-                    CompositeDecoder.DECODE_DONE -> break@loop // https://youtrack.jetbrains.com/issue/KT-42262
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            DatePeriod(years, months, days)
-        }
-
-    override fun serialize(encoder: Encoder, value: DatePeriod) {
-        encoder.encodeStructure(descriptor) {
-            with(value) {
-                if (years != 0) encodeIntElement(DateTimePeriodComponentSerializer.descriptor, 0, years)
-                if (months != 0) encodeIntElement(DateTimePeriodComponentSerializer.descriptor, 1, months)
-                if (days != 0) encodeIntElement(DateTimePeriodComponentSerializer.descriptor, 2, days)
-            }
-        }
-    }
-
-}
-
-object DatePeriodISO8601Serializer: KSerializer<DatePeriod> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("DatePeriod", PrimitiveKind.STRING)
-
-    // TODO: consider whether should fail when parsing "P1YT0H0M0.0S"
-    override fun deserialize(decoder: Decoder): DatePeriod =
-        when (val period = DateTimePeriod.parse(decoder.decodeString())) {
-            is DatePeriod -> period
-            else -> throw IllegalArgumentException("$period is not a date-based period")
-        }
-
-    override fun serialize(encoder: Encoder, value: DatePeriod) {
-        encoder.encodeString(value.toString())
-    }
-
-}
 
 @Serializable(with = DatePeriodISO8601Serializer::class)
 class DatePeriod internal constructor(
