+++ core/common/src/DayOfWeek.kt
@@ -5,23 +5,7 @@
 
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-import kotlinx.serialization.internal.*
 import kotlin.native.concurrent.*
-
-@Suppress("INVISIBLE_MEMBER")
-object DayOfWeekSerializer: KSerializer<DayOfWeek> {
-    private val impl = EnumSerializer("Month", DayOfWeek.values())
-
-    override val descriptor: SerialDescriptor
-        get() = impl.descriptor
-
-    override fun deserialize(decoder: Decoder): DayOfWeek = impl.deserialize(decoder)
-
-    override fun serialize(encoder: Encoder, value: DayOfWeek) = impl.serialize(encoder, value)
-}
 
 public expect enum class DayOfWeek {
     MONDAY,
