+++ core/common/src/LocalDate.kt
@@ -5,66 +5,9 @@
 
 package kotlinx.datetime
 
+import kotlinx.datetime.serializers.LocalDateISO8601Serializer
+import kotlinx.serialization.Serializable
 import kotlin.time.*
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-
-object LocalDateISO8601Serializer: KSerializer<LocalDate> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("LocalDate", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): LocalDate =
-        LocalDate.parse(decoder.decodeString())
-
-    override fun serialize(encoder: Encoder, value: LocalDate) {
-        encoder.encodeString(value.toString())
-    }
-
-}
-
-object LocalDateComponentSerializer: KSerializer<LocalDate> {
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("LocalDate") {
-            element<Int>("year")
-            element<Short>("month")
-            element<Short>("day")
-        }
-
-    @Suppress("INVISIBLE_MEMBER") // to be able to throw `MissingFieldException`
-    override fun deserialize(decoder: Decoder): LocalDate =
-        decoder.decodeStructure(descriptor) {
-            var year: Int? = null
-            var month: Short? = null
-            var day: Short? = null
-            loop@while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> year = decodeIntElement(descriptor, 0)
-                    1 -> month = decodeShortElement(descriptor, 1)
-                    2 -> day = decodeShortElement(descriptor, 2)
-                    CompositeDecoder.DECODE_DONE -> break@loop // https://youtrack.jetbrains.com/issue/KT-42262
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            if (year == null) throw MissingFieldException("year")
-            if (month == null) throw MissingFieldException("month")
-            if (day == null) throw MissingFieldException("day")
-            LocalDate(year, month.toInt(), day.toInt())
-        }
-
-    override fun serialize(encoder: Encoder, value: LocalDate) {
-        encoder.encodeStructure(descriptor) {
-            encodeIntElement(descriptor, 0, value.year)
-            encodeShortElement(descriptor, 1, value.monthNumber.toShort())
-            encodeShortElement(descriptor, 2, value.dayOfMonth.toShort())
-        }
-    }
-
-}
-
-expect object LocalDateLongSerializer: KSerializer<LocalDate>
 
 @Serializable(with = LocalDateISO8601Serializer::class)
 public expect class LocalDate : Comparable<LocalDate> {
