+++ core/common/src/LocalDateTime.kt
@@ -5,87 +5,8 @@
 
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-
-object LocalDateTimeISO8601Serializer: KSerializer<LocalDateTime> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("LocalDateTime", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): LocalDateTime =
-        LocalDateTime.parse(decoder.decodeString())
-
-    override fun serialize(encoder: Encoder, value: LocalDateTime) {
-        encoder.encodeString(value.toString())
-    }
-
-}
-
-object LocalDateTimeComponentSerializer: KSerializer<LocalDateTime> {
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("LocalDateTime") {
-            element<Int>("year")
-            element<Short>("month")
-            element<Short>("day")
-            element<Short>("hour")
-            element<Short>("minute")
-            element<Short>("second", isOptional = true)
-            element<Int>("nanosecond", isOptional = true)
-        }
-
-    @Suppress("INVISIBLE_MEMBER") // to be able to throw `MissingFieldException`
-    override fun deserialize(decoder: Decoder): LocalDateTime =
-        decoder.decodeStructure(descriptor) {
-            var year: Int? = null
-            var month: Short? = null
-            var day: Short? = null
-            var hour: Short? = null
-            var minute: Short? = null
-            var second: Short = 0
-            var nanosecond = 0
-            loop@while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> year = decodeIntElement(descriptor, 0)
-                    1 -> month = decodeShortElement(descriptor, 1)
-                    2 -> day = decodeShortElement(descriptor, 2)
-                    3 -> hour = decodeShortElement(descriptor, 3)
-                    4 -> minute = decodeShortElement(descriptor, 4)
-                    5 -> second = decodeShortElement(descriptor, 5)
-                    6 -> nanosecond = decodeIntElement(descriptor, 6)
-                    CompositeDecoder.DECODE_DONE -> break@loop // https://youtrack.jetbrains.com/issue/KT-42262
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            if (year == null) throw MissingFieldException("year")
-            if (month == null) throw MissingFieldException("month")
-            if (day == null) throw MissingFieldException("day")
-            if (hour == null) throw MissingFieldException("hour")
-            if (minute == null) throw MissingFieldException("minute")
-            LocalDateTime(year, month.toInt(), day.toInt(), hour.toInt(), minute.toInt(), second.toInt(), nanosecond)
-        }
-
-    override fun serialize(encoder: Encoder, value: LocalDateTime) {
-        encoder.encodeStructure(descriptor) {
-            encodeIntElement(descriptor, 0, value.year)
-            encodeShortElement(descriptor, 1, value.monthNumber.toShort())
-            encodeShortElement(descriptor, 2, value.dayOfMonth.toShort())
-            encodeShortElement(descriptor, 3, value.hour.toShort())
-            encodeShortElement(descriptor, 4, value.minute.toShort())
-            if (value.second != 0 || value.nanosecond != 0) {
-                encodeShortElement(descriptor, 5, value.second.toShort())
-                if (value.nanosecond != 0) {
-                    encodeIntElement(descriptor, 6, value.nanosecond)
-                }
-            }
-        }
-    }
-
-}
-
-expect object LocalDateTimeCompactSerializer: KSerializer<LocalDateTime>
+import kotlinx.datetime.serializers.LocalDateTimeISO8601Serializer
+import kotlinx.serialization.Serializable
 
 @Serializable(with = LocalDateTimeISO8601Serializer::class)
 public expect class LocalDateTime : Comparable<LocalDateTime> {
