+++ core/common/src/TimeZone.kt
@@ -8,42 +8,9 @@
 
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-
-object TimeZoneSerializer: KSerializer<TimeZone> {
-
-    override val descriptor: SerialDescriptor
-        get() = PrimitiveSerialDescriptor("TimeZone", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): TimeZone = TimeZone.of(decoder.decodeString())
-
-    override fun serialize(encoder: Encoder, value: TimeZone) {
-        encoder.encodeString(value.id)
-    }
-
-}
-
-object ZoneOffsetSerializer: KSerializer<ZoneOffset> {
-
-    override val descriptor: SerialDescriptor
-        get() = PrimitiveSerialDescriptor("ZoneOffset", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): ZoneOffset {
-        val zone = TimeZone.of(decoder.decodeString())
-        if (zone is ZoneOffset) {
-            return zone
-        } else {
-            throw SerializationException("Timezone identifier '$zone' does not correspond to a fixed-offset timezone")
-        }
-    }
-
-    override fun serialize(encoder: Encoder, value: ZoneOffset) {
-        encoder.encodeString(value.id)
-    }
-
-}
+import kotlinx.datetime.serializers.TimeZoneSerializer
+import kotlinx.datetime.serializers.ZoneOffsetSerializer
+import kotlinx.serialization.Serializable
 
 @Serializable(with = TimeZoneSerializer::class)
 public expect open class TimeZone {
