+++ core/js/src/Instant.kt
@@ -14,6 +14,7 @@ import kotlinx.datetime.internal.JSJoda.Instant as jtInstant
 import kotlinx.datetime.internal.JSJoda.Duration as jtDuration
 import kotlinx.datetime.internal.JSJoda.Clock as jtClock
 import kotlinx.datetime.internal.JSJoda.ChronoUnit
+import kotlinx.datetime.serializers.InstantISO8601Serializer
 import kotlinx.serialization.Serializable
 import kotlin.math.truncate
