+++ core/js/src/LocalDateTime.kt
@@ -4,47 +4,9 @@
  */
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
+import kotlinx.datetime.serializers.LocalDateTimeISO8601Serializer
+import kotlinx.serialization.Serializable
 import kotlinx.datetime.internal.JSJoda.LocalDateTime as jtLocalDateTime
-import kotlinx.datetime.internal.JSJoda.LocalDate as jtLocalDate
-import kotlinx.datetime.internal.JSJoda.LocalTime as jtLocalTime
-
-actual object LocalDateTimeCompactSerializer: KSerializer<LocalDateTime> {
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("LocalDateTime") {
-            element<Long>("epochDay")
-            element<Long>("nanoOfDay")
-        }
-
-    @Suppress("INVISIBLE_MEMBER") // to be able to throw `MissingFieldException`
-    override fun deserialize(decoder: Decoder): LocalDateTime =
-        decoder.decodeStructure(descriptor) {
-            var epochDay: Long? = null
-            var nanoOfDay: Long? = null
-            while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> epochDay = decodeLongElement(descriptor, 0)
-                    1 -> nanoOfDay = decodeLongElement(descriptor, 1)
-                    CompositeDecoder.DECODE_DONE -> break
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            if (epochDay == null) throw MissingFieldException("epochDay")
-            if (nanoOfDay == null) throw MissingFieldException("nanoOfDay")
-            LocalDateTime(jtLocalDateTime.of(jtLocalDate.ofEpochDay(epochDay), jtLocalTime.ofNanoOfDay(nanoOfDay)))
-        }
-
-    override fun serialize(encoder: Encoder, value: LocalDateTime) {
-        encoder.encodeStructure(descriptor) {
-            encodeLongElement(descriptor, 0, value.date.value.toEpochDay().toLong())
-            encodeLongElement(descriptor, 1, value.value.toLocalTime().toNanoOfDay().toLong())
-        }
-    }
-
-}
 
 @Serializable(with = LocalDateTimeISO8601Serializer::class)
 public actual class LocalDateTime internal constructor(internal val value: jtLocalDateTime) : Comparable<LocalDateTime> {
