+++ core/jvm/src/LocalDate.kt
@@ -5,27 +5,12 @@
 @file:JvmName("LocalDateJvmKt")
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
+import kotlinx.datetime.serializers.LocalDateISO8601Serializer
+import kotlinx.serialization.Serializable
 import java.time.DateTimeException
 import java.time.format.DateTimeParseException
 import java.time.temporal.ChronoUnit
 import java.time.LocalDate as jtLocalDate
-
-actual object LocalDateLongSerializer: KSerializer<LocalDate> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("LocalDate", PrimitiveKind.LONG)
-
-    override fun deserialize(decoder: Decoder): LocalDate =
-        LocalDate(jtLocalDate.ofEpochDay(decoder.decodeLong()))
-
-    override fun serialize(encoder: Encoder, value: LocalDate) {
-        encoder.encodeLong(value.value.toEpochDay())
-    }
-
-}
 
 @Serializable(with = LocalDateISO8601Serializer::class)
 public actual class LocalDate internal constructor(internal val value: jtLocalDate) : Comparable<LocalDate> {
