+++ core/jvm/src/LocalDateTime.kt
@@ -5,53 +5,14 @@
 @file:JvmName("LocalDateTimeJvmKt")
 package kotlinx.datetime
 
-import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
+import kotlinx.datetime.serializers.LocalDateTimeISO8601Serializer
+import kotlinx.serialization.Serializable
 import java.time.DateTimeException
 import java.time.format.DateTimeParseException
 import java.time.LocalDateTime as jtLocalDateTime
-import java.time.LocalDate as jtLocalDate
-import java.time.LocalTime as jtLocalTime
-import kotlinx.serialization.internal.*
 
 public actual typealias Month = java.time.Month
 public actual typealias DayOfWeek = java.time.DayOfWeek
-
-actual object LocalDateTimeCompactSerializer: KSerializer<LocalDateTime> {
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("LocalDateTime") {
-            element<Long>("epochDay")
-            element<Long>("nanoOfDay")
-        }
-
-    @Suppress("INVISIBLE_MEMBER") // to be able to throw `MissingFieldException`
-    override fun deserialize(decoder: Decoder): LocalDateTime =
-        decoder.decodeStructure(descriptor) {
-            var epochDay: Long? = null
-            var nanoOfDay: Long? = null
-            while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> epochDay = decodeLongElement(descriptor, 0)
-                    1 -> nanoOfDay = decodeLongElement(descriptor, 1)
-                    CompositeDecoder.DECODE_DONE -> break
-                    else -> error("Unexpected index: $index")
-                }
-            }
-            if (epochDay == null) throw MissingFieldException("epochDay")
-            if (nanoOfDay == null) throw MissingFieldException("nanoOfDay")
-            LocalDateTime(jtLocalDateTime.of(jtLocalDate.ofEpochDay(epochDay), jtLocalTime.ofNanoOfDay(nanoOfDay)))
-        }
-
-    override fun serialize(encoder: Encoder, value: LocalDateTime) {
-        encoder.encodeStructure(descriptor) {
-            encodeLongElement(descriptor, 0, value.date.value.toEpochDay())
-            encodeLongElement(descriptor, 1, value.value.toLocalTime().toNanoOfDay())
-        }
-    }
-
-}
 
 @Serializable(with = LocalDateTimeISO8601Serializer::class)
 public actual class LocalDateTime internal constructor(internal val value: jtLocalDateTime) : Comparable<LocalDateTime> {
