+++ core/native/src/Instant.kt
@@ -8,7 +8,8 @@
 
 package kotlinx.datetime
 
-import kotlinx.serialization.*
+import kotlinx.datetime.serializers.InstantISO8601Serializer
+import kotlinx.serialization.Serializable
 import kotlin.math.*
 import kotlin.time.*
