+++ core/native/src/serializers/LocalDateSerializers.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2019-2021 JetBrains s.r.o.
+ * Use of this source code is governed by the Apache 2.0 License that can be found in the LICENSE.txt file.
+ */
+
+package kotlinx.datetime.serializers
+
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.clampToInt
+import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
+import kotlinx.serialization.encoding.*
+
+actual object LocalDateLongSerializer: KSerializer<LocalDate> {
+
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("LocalDate", PrimitiveKind.LONG)
+
+    override fun deserialize(decoder: Decoder): LocalDate =
+        LocalDate.ofEpochDay(decoder.decodeLong().clampToInt())
+
+    override fun serialize(encoder: Encoder, value: LocalDate) {
+        encoder.encodeLong(value.toEpochDay().toLong())
+    }
+
+}
