+++ serialization/common/test/LocalDateTimeSerializationTest.kt
@@ -6,6 +6,7 @@
 package kotlinx.datetime.serialization.test
 
 import kotlinx.datetime.*
+import kotlinx.datetime.serializers.*
 import kotlinx.serialization.json.*
 import kotlin.test.*
